+++ src/Expression/Builder/ExpressionBuilderInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Expression;
+namespace Yiisoft\Db\Expression\Builder;
+
+use Yiisoft\Db\Expression\ExpressionInterface as T;
 
 /**
  * This interface defines the methods to build database expressions, such as conditions for a SELECT statement or values
@@ -16,7 +18,7 @@ namespace Yiisoft\Db\Expression;
  *
  * @see ExpressionInterface
  *
- * @template T as ExpressionInterface
+ * @template T as T
  */
 interface ExpressionBuilderInterface
 {
@@ -24,12 +26,12 @@ interface ExpressionBuilderInterface
      * Method builds the raw SQL from the expression that will not be additionally
      * escaped or quoted.
      *
-     * @param ExpressionInterface $expression The expression to be built.
+     * @param T $expression The expression to be built.
      * @param array $params The binding parameters.
      * @throws \InvalidArgumentException If builder can't handle expression passed.
      * @return string The raw SQL that will not be additionally escaped or quoted.
      *
      * @psalm-param T $expression
      */
-    public function build(ExpressionInterface $expression, array &$params = []): string;
+    public function build(T $expression, array &$params = []): string;
 }
