+++ src/QueryBuilder/AbstractDQLQueryBuilder.php
@@ -4,28 +4,28 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\QueryBuilder;
 
-use Yiisoft\Db\Command\Param;
-use Yiisoft\Db\Command\ParamBuilder;
+use Yiisoft\Db\Expression\Param;
+use Yiisoft\Db\Expression\Builder\ParamBuilder;
 use Yiisoft\Db\Exception\Exception;
 use InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\ArrayExpression;
-use Yiisoft\Db\Expression\ArrayExpressionBuilder;
+use Yiisoft\Db\Expression\Builder\ArrayExpressionBuilder;
 use Yiisoft\Db\Expression\ColumnName;
-use Yiisoft\Db\Expression\ColumnNameBuilder;
+use Yiisoft\Db\Expression\Builder\ColumnNameBuilder;
 use Yiisoft\Db\Expression\Expression;
-use Yiisoft\Db\Expression\ExpressionBuilder;
-use Yiisoft\Db\Expression\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\Builder\ExpressionBuilder;
+use Yiisoft\Db\Expression\Builder\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Expression\JsonExpression;
-use Yiisoft\Db\Expression\JsonExpressionBuilder;
+use Yiisoft\Db\Expression\Builder\JsonExpressionBuilder;
 use Yiisoft\Db\Expression\CaseExpression;
-use Yiisoft\Db\Expression\CaseExpressionBuilder;
+use Yiisoft\Db\Expression\Builder\CaseExpressionBuilder;
 use Yiisoft\Db\Expression\StructuredExpression;
-use Yiisoft\Db\Expression\StructuredExpressionBuilder;
+use Yiisoft\Db\Expression\Builder\StructuredExpressionBuilder;
 use Yiisoft\Db\Expression\Value;
-use Yiisoft\Db\Expression\ValueBuilder;
+use Yiisoft\Db\Expression\Builder\ValueBuilder;
 use Yiisoft\Db\QueryBuilder\Condition\ConditionInterface;
 use Yiisoft\Db\QueryBuilder\Condition\Simple;
 use Yiisoft\Db\Query\Query;
