+++ src/Schema/Column/BinaryColumn.php
@@ -7,7 +7,7 @@ namespace Yiisoft\Db\Schema\Column;
 use BackedEnum;
 use PDO;
 use Stringable;
-use Yiisoft\Db\Command\Param;
+use Yiisoft\Db\Expression\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Constant\GettypeResult;
