+++ tests/Db/QueryBuilder/Condition/Builder/EqualsBuilderTest.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Tests\Db\QueryBuilder\Condition\Builder;
 
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Db\Command\Param;
+use Yiisoft\Db\Expression\Param;
 use Yiisoft\Db\Constant\DataType;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\QueryBuilder\Condition\Builder\EqualsBuilder;
