+++ tests/Provider/QueryBuilderProvider.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\Tests\Provider;
 
 use ArrayIterator;
 use Yiisoft\Db\Constant\DataType;
-use Yiisoft\Db\Command\Param;
+use Yiisoft\Db\Expression\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constant\IndexType;
 use Yiisoft\Db\Constant\PseudoType;
